refactor: make into command
